+++ dashboard/src/components/ui/Avatar.tsx
@@ -59,6 +59,7 @@ export function Avatar({
     <Box
       style={Object.assign(style, { backgroundImage: `url(${avatarUrl})` })}
       className={classes}
+      aria-label="Avatar"
       {...rest}
     />
   );