+++ dashboard/src/pages/signin/email.tsx
@@ -8,7 +8,7 @@ import Text from '@/ui/v2/Text';
 import { getToastStyleProps } from '@/utils/settings/settingsConstants';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { styled } from '@mui/material';
-import { useResetPassword } from '@nhost/nextjs';
+import { useSignInEmailPassword } from '@nhost/nextjs';
 import type { ReactElement } from 'react';
 import { useEffect } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
@@ -17,9 +17,10 @@ import * as Yup from 'yup';
 
 const validationSchema = Yup.object({
   email: Yup.string().label('Email').email().required(),
+  password: Yup.string().label('Password').required(),
 });
 
-export type ResetPasswordFormValues = Yup.InferType<typeof validationSchema>;
+export type EmailSignUpFormValues = Yup.InferType<typeof validationSchema>;
 
 const StyledInput = styled(Input)({
   backgroundColor: 'transparent',
@@ -28,18 +29,19 @@ const StyledInput = styled(Input)({
   },
 });
 
-export default function ResetPasswordPage() {
-  const { resetPassword, error, isSent } = useResetPassword();
+export default function EmailSignUpPage() {
+  const { signInEmailPassword, error } = useSignInEmailPassword();
 
-  const form = useForm<ResetPasswordFormValues>({
+  const form = useForm<EmailSignUpFormValues>({
     reValidateMode: 'onSubmit',
     defaultValues: {
       email: '',
+      password: '',
     },
     resolver: yupResolver(validationSchema),
   });
 
-  const { register, formState, getValues } = form;
+  const { register, formState } = form;
 
   useEffect(() => {
     if (!error) {
@@ -52,24 +54,15 @@ export default function ResetPasswordPage() {
     );
   }, [error]);
 
-  async function handleSubmit({ email }: ResetPasswordFormValues) {
+  async function handleSubmit({ email, password }: EmailSignUpFormValues) {
     try {
-      await resetPassword(email);
+      await signInEmailPassword(email, password);
     } catch {
       toast.error(
-        'An error occurred while signing up. Please try again.',
+        'An error occurred while signing in. Please try again.',
         getToastStyleProps(),
       );
     }
-  }
-
-  if (isSent) {
-    return (
-      <div className="text-center">
-        We&apos;ve sent a temporary link to reset your password. Check your
-        inbox at {getValues('email')}.
-      </div>
-    );
   }
 
   return (
@@ -79,7 +72,7 @@ export default function ResetPasswordPage() {
         component="h1"
         className="text-center text-3.5xl font-semibold lg:text-4.5xl"
       >
-        Reset Password
+        Sign In
       </Text>
 
       <Box className="grid grid-flow-row gap-4 rounded-md border bg-transparent p-6 lg:p-12">
@@ -90,42 +83,73 @@ export default function ResetPasswordPage() {
           >
             <StyledInput
               {...register('email')}
-              type="email"
               id="email"
-              label="Email"
               placeholder="Email"
-              fullWidth
-              autoFocus
               inputProps={{ min: 2, max: 128 }}
+              spellCheck="false"
+              autoCapitalize="none"
+              type="email"
+              label="Email"
+              hideEmptyHelperText
+              fullWidth
               error={!!formState.errors.email}
               helperText={formState.errors.email?.message}
+              autoFocus
             />
+
+            <StyledInput
+              {...register('password')}
+              id="password"
+              placeholder="Password"
+              inputProps={{ min: 2, max: 128 }}
+              spellCheck="false"
+              autoCapitalize="none"
+              type="password"
+              label="Password"
+              hideEmptyHelperText
+              fullWidth
+              error={!!formState.errors.password}
+              helperText={formState.errors.password?.message}
+            />
+
+            <NavLink
+              href="/reset-password"
+              color="white"
+              className="justify-self-start font-semibold"
+            >
+              Forgot password?
+            </NavLink>
 
             <Button
               className="!bg-white !text-black disabled:!text-black disabled:!text-opacity-60"
               size="large"
-              type="submit"
               disabled={formState.isSubmitting}
               loading={formState.isSubmitting}
+              type="submit"
             >
-              Send Reset Instructions
+              Sign In
             </Button>
+
+            <Text color="secondary" className="text-center">
+              or{' '}
+              <NavLink color="white" className="font-semibold" href="/signin">
+                sign in with GitHub
+              </NavLink>
+            </Text>
           </Form>
         </FormProvider>
       </Box>
 
       <Text color="secondary" className="text-center text-base lg:text-lg">
-        Is your password okay?{' '}
-        <NavLink href="/signin/email" color="white" className="font-medium">
-          Sign In
+        Don&apos;t have an account?{' '}
+        <NavLink href="/signup" color="white">
+          Sign Up
         </NavLink>
       </Text>
     </>
   );
 }
 
-ResetPasswordPage.getLayout = function getLayout(page: ReactElement) {
-  return (
-    <UnauthenticatedLayout title="Reset Password">{page}</UnauthenticatedLayout>
-  );
+EmailSignUpPage.getLayout = function getLayout(page: ReactElement) {
+  return <UnauthenticatedLayout title="Sign In">{page}</UnauthenticatedLayout>;
 };